requirements.txt
telethon
aiohttp
ruamel.yaml
bs4
tld